+++ R/pval_plot.R
@@ -14,7 +14,7 @@
 #' }
 #'
 #' @importFrom raster cut raster
-#' @importFrom spatstat as.im
+#' @importFrom spatstat.geom as.im
 #' @export
 #'
 #' @keywords internal
@@ -27,7 +27,7 @@ pval_plot <- function(input,
     stop("The 'input' argument must be an object of class 'im'")
   }
 
-  out <- raster::raster(spatstat::as.im(input))  # create raster
+  out <- raster::raster(spatstat.geom::as.im(input))  # create raster
   out <- raster::cut(out,
                      breaks = c(-Inf, alpha / 2, 1 - alpha / 2, Inf))
   return(out)