+++ NEWS.md
@@ -1,5 +1,9 @@
 # gateR (development version)
 
+# gateR v0.1.7
+  * Updated {spatstat} dependency to new packages based on feedback from the Spatstat Team (Adrian Baddeley and Ege Rubak). {spatstat.geom} replaces {spatstat} in Imports
+  * Added additional multiple testing corrections, including False Discovery Rate, spatially dependent Sidak correction, and independent Sidak correction
+
 # gateR v0.1.6
   * Updated URLs in gateR-package.Rd
   * Updated year in DESCRIPTION